Fix Outlook border-left color
Apparently the Outlook native app doesn't support border-color:inherit,
and I'd forgotten to test this when switching the styling to use it.
Instead, place the border on the whole block object to avoid needing to
inherit anything.

Add \n to the HTML output to avoid very long lines in HTML. This wastes
a bit of space, but prevents lines from being split in unexpected places,
which breaks the markup.

GMail Android in dark theme mode changes the color for border-left when
it's full yellow. Switch to a slightly different color that isn't
affected!

Fixes #1