make server command interactions synchronous
This ensures that server-side handling of block start/end requests has
been fully completed before clients continue to emit logs that are
intended to be/not-be part of the prior/next block.

This relates to #2, but only solves the part of the issue described in
the initial description, not the follow-up comment.